Get some popcorn, this is a LONG commit. This commit marks the first working version of the: Michael Xiong League of Legends Champion Spell Efficiency Calculator!
Things to note:
I deleted champion.py (it was useless)
An explanation of all of the things changed since the last commit, and
an explanation of how this entire thing works can be found in README.
Wonder why it's called README.... Seriously, read it. I spent a while
writing it.

TLDR: I FINISHED THE FIRST WORKING COPY, README EXPLAINS IT